# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Simple server in order to create and mantain CGA Survey"
  version: "1.0.0"
  title: "CGA Survey"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "a@a.com"
  license:
    name: "MIT"
    url: "https://github.com/TrentoCrowdAI/CGA/blob/master/LICENSE"
host: "cga-mock-server.herokuapp.com"
basePath: "/v1"
...
schemes:
- "https"
paths:
...
  /surveyItems:
    get:
      tags:
      - "surveyItem"
      summary: "Get all surveyItem"
      description: "Performed by a user in order to get all surveyItem"
      operationId: "getAllSurveyItems"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/surveyItem"
        400:
          description: "Bad request"
      x-swagger-router-controller: "SurveyItem"
    post:
      tags:
      - "surveyItem"
      summary: "Add a new SurveyItem"
      description: "Performed by a user when he wants to create a new SurveyItem"
      operationId: "addSurveyItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "SurveyItem"
        description: "SurveyItem object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/SurveyItem"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyItem"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "SurveyItem"
  /surveyItems/{id}:
    get:
      tags:
      - "surveyItem"
      summary: "Get a SurveyItem by the id"
      description: "Performed by users in order to get specific SurveyItem"
      operationId: "getSurveyItem"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyItem id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/surveyItem"
        400:
          description: "Invalid id"
        404:
          description: "SurveyItem not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyItem"
    put:
      tags:
      - "surveyItem"
      summary: "Update an existing SurveyItem"
      description: "Performed by a user in order to update the surveyItem"
      operationId: "updateSurveyItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyItem id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "SurveyItem object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_4"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyItem"
        400:
          description: "Invalid id"
        404:
          description: "SurveyItem not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyItem"
    delete:
      tags:
      - "surveyItem"
      summary: "Delete a SurveyItem by id"
      description: "Performed by the user in order to delete the SurveyItem from the\
        \ system"
      operationId: "deleteSurveyItem"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyItem id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyItem"
        400:
          description: "Invalid id"
        404:
          description: "SurveyItem not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyItem"
  /surveyImages:
    get:
      tags:
      - "surveyImage"
      summary: "Get all SurveyImage"
      description: "Performed by a user in order to get all SurveyImage"
      operationId: "getAllSurveyImages"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/surveyImage"
        400:
          description: "Bad request"
      x-swagger-router-controller: "SurveyImage"
    post:
      tags:
      - "surveyImage"
      summary: "Add a new SurveyImage"
      description: "Performed by a user when he wants to create a new SurveyImage"
      operationId: "addSurveyImage"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "SurveyImage"
        description: "SurveyImage object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/SurveyImage"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyImage"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "SurveyImage"
  /surveyImages/{id}:
    get:
      tags:
      - "surveyImage"
      summary: "Get a SurveyItem by the id"
      description: "Performed by users in order to get specific SurveyImage"
      operationId: "getSurveyImage"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyImage id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/surveyImage"
        400:
          description: "Invalid id"
        404:
          description: "SurveyImage not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyImage"
    put:
      tags:
      - "surveyImage"
      summary: "Update an existing SurveyImage"
      description: "Performed by a user in order to update the SurveyImage"
      operationId: "updateSurveyImage"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyImage id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "SurveyImage object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_5"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyImage"
        400:
          description: "Invalid id"
        404:
          description: "SurveyImage not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyImage"
    delete:
      tags:
      - "surveyImage"
      summary: "Delete a SurveyImage by id"
      description: "Performed by the user in order to delete the SurveyImage from\
        \ the system"
      operationId: "deleteSurveyImage"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyImage id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyImage"
        400:
          description: "Invalid id"
        404:
          description: "SurveyImage not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyImage"
...
  /validatorItems:
    get:
      tags:
      - "validatorItem"
      summary: "Get all validatorItem"
      description: "Performed by a user in order to get all ValidatorItem"
      operationId: "getAllValidatorItems"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/validatorItem"
        400:
          description: "Bad request"
      x-swagger-router-controller: "ValidatorItem"
    post:
      tags:
      - "validatorItem"
      summary: "Add a new ValidatorItem"
      description: "Operation to call in order to create a new ValidatorItem in the\
        \ system"
      operationId: "addValidatorItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "validatorItem"
        description: "ValidatorItem object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/validatorItem_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/validatorItem"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "ValidatorItem"
  /validatorItems/{id}:
    get:
      tags:
      - "validatorItem"
      summary: "Get a ValidatorItem by the id"
      description: "Performed by users in order to get specific ValidatorItem"
      operationId: "getValidatorItem"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ValidatorItem id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/validatorItem"
        400:
          description: "Invalid id"
        404:
          description: "ValidatorItem not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "ValidatorItem"
    put:
      tags:
      - "validatorItem"
      summary: "Update an existing ValidatorItem"
      description: "Performed by a user in order to update the ValidatorItem"
      operationId: "updateValidatorItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ValidatorItem id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Subject object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_8"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/validatorItem"
        400:
          description: "Invalid id"
        404:
          description: "ValidatorItem not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "ValidatorItem"
    delete:
      tags:
      - "validatorItem"
      summary: "Delete a ValidatorItem by id"
      description: "Performed by the user in order to delete the ValidatorItem from\
        \ the system"
      operationId: "deleteValidatorItem"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ValidatorItem id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/validatorItem"
        400:
          description: "Invalid id"
        404:
          description: "ValidatorItem not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "ValidatorItem"
...
  /surveyItemOptions:
    get:
      tags:
      - "surveyItemOption"
      summary: "Get all SurveyItemOption"
      description: "Performed by a user in order to get all SurveyItemOption"
      operationId: "getAllSurveyItemOptions"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/surveyItemOption"
        400:
          description: "Bad request"
      x-swagger-router-controller: "SurveyItemOption"
    post:
      tags:
      - "surveyItemOption"
      summary: "Add a new SurveyItemOption"
      description: "Performed by a user when he wants to create a new SurveyItemOption"
      operationId: "addSurveyItemOption"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "surveyItemOption"
        description: "SurveyItemOption object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/surveyItemOption_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyItemOption"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "SurveyItemOption"
  /surveyItemOptions/{id}:
    get:
      tags:
      - "surveyItemOption"
      summary: "Get a SurveyItemOption by the id"
      description: "Performed by users in order to get specific SurveyItemOption"
      operationId: "getSurveyItemOption"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyItemOption id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/surveyItemOption"
        400:
          description: "Invalid id"
        404:
          description: "SurveyItemOption not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyItemOption"
    put:
      tags:
      - "surveyItemOption"
      summary: "Update an existing SurveyItemOption"
      description: "Performed by a user in order to update the SurveyItemOption"
      operationId: "updateSurveyItemOption"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyItemOption id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "SurveyItemOption object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_10"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyItemOption"
        400:
          description: "Invalid id"
        404:
          description: "SurveyItemOption not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyItemOption"
    delete:
      tags:
      - "surveyItemOption"
      summary: "Delete a SurveyItemOption by id"
      description: "Performed by the user in order to delete the SurveyItemOption\
        \ from the system"
      operationId: "deleteSurveyItemOption"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyItemOption id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyItemOption"
        400:
          description: "Invalid id"
        404:
          description: "SurveyItemOption not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyItemOption"
...
definitions:
...
  surveyItem:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the SurveyItem"
      description:
        type: "string"
        description: "Description for the item"
      labels:
        type: "array"
        description: "Objects that allow the surveyItem to be displayed in different\
          \ languages"
        items:
          $ref: "#/definitions/label"
      type:
        type: "string"
        description: "Name that identify the type of the surveyItemOption"
      options:
        type: "array"
        description: "Option that the user can select"
        items:
          $ref: "#/definitions/surveyItemOption"
      rules:
        type: "array"
        description: "Rules in order to apply consistency checks"
        items:
          $ref: "#/definitions/rule"
      validators:
        type: "array"
        description: "Object in order to validate the user answers"
        items:
          $ref: "#/definitions/validatorItem"
      flowConstraints:
        type: "array"
        description: "FlowConstraints to apply"
        items:
          $ref: "#/definitions/flowConstraint"
    example:
      validators:
      - id: 2
        type: "type"
        value: "value"
      - id: 2
        type: "type"
        value: "value"
      options:
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      description: "description"
      rules:
      - item: 3
        condition: "condition"
        id: 9
      - item: 3
        condition: "condition"
        id: 9
      id: 5
      type: "type"
      flowConstraints:
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      labels:
      - values: "values"
        language: "language"
        id: 2
      - values: "values"
        language: "language"
        id: 2
  surveyImage:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the SurveyItem"
      description:
        type: "string"
        description: "Description for the item"
      link:
        type: "string"
        description: "Link to the image file"
      labels:
        type: "array"
        description: "Objects that allow the surveyItem to be displayed in different\
          \ languages"
        items:
          $ref: "#/definitions/label"
      type:
        type: "string"
        description: "Name that identify the type of the surveyItemOption"
      options:
        type: "array"
        description: "Option that the user can select"
        items:
          $ref: "#/definitions/surveyItemOption"
      rules:
        type: "array"
        description: "Rules in order to apply consistency checks"
        items:
          $ref: "#/definitions/rule"
      validators:
        type: "array"
        description: "Object in order to validate the user answers"
        items:
          $ref: "#/definitions/validatorItem"
      flowConstraints:
        type: "array"
        description: "FlowConstraints to apply"
        items:
          $ref: "#/definitions/flowConstraint"
    example:
      validators:
      - id: 2
        type: "type"
        value: "value"
      - id: 2
        type: "type"
        value: "value"
      link: "link"
      options:
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      description: "description"
      rules:
      - item: 3
        condition: "condition"
        id: 9
      - item: 3
        condition: "condition"
        id: 9
      id: 0
      type: "type"
      flowConstraints:
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      labels:
      - values: "values"
        language: "language"
        id: 2
      - values: "values"
        language: "language"
        id: 2
...
  validatorItem:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the validatorItem in the system"
      type:
        type: "string"
        description: "Type of the validation value"
      value:
        type: "string"
        description: "Value in order to validate the data"
    example:
      id: 2
      type: "type"
      value: "value"
...
  SurveyItem:
    type: "object"
    properties:
      description:
        type: "string"
        description: "Description for the item"
      labels:
        type: "array"
        description: "Objects that allow the surveyItem to be displayed in different\
          \ languages"
        items:
          $ref: "#/definitions/label"
      type:
        type: "string"
        description: "Name that identify the type of the surveyItemOption"
      options:
        type: "array"
        description: "Option that the user can select"
        items:
          $ref: "#/definitions/surveyItemOption"
      rules:
        type: "array"
        description: "Rules in order to apply consistency checks"
        items:
          $ref: "#/definitions/rule"
      validators:
        type: "array"
        description: "Object in order to validate the user answers"
        items:
          $ref: "#/definitions/validatorItem"
      flowConstraints:
        type: "array"
        description: "FlowConstraints to apply"
        items:
          $ref: "#/definitions/flowConstraint"
    example:
      validators:
      - id: 2
        type: "type"
        value: "value"
      - id: 2
        type: "type"
        value: "value"
      options:
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      description: "description"
      rules:
      - item: 3
        condition: "condition"
        id: 9
      - item: 3
        condition: "condition"
        id: 9
      type: "type"
      flowConstraints:
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      labels:
      - values: "values"
        language: "language"
        id: 2
      - values: "values"
        language: "language"
        id: 2
  body_4:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the SurveyItem"
      description:
        type: "string"
        description: "Description for the item"
      labels:
        type: "array"
        description: "Objects that allow the surveyItem to be displayed in different\
          \ languages"
        items:
          $ref: "#/definitions/label"
      type:
        type: "string"
        description: "Name that identify the type of the surveyItemOption"
      options:
        type: "array"
        description: "Option that the user can select"
        items:
          $ref: "#/definitions/surveyItemOption"
      rules:
        type: "array"
        description: "Rules in order to apply consistency checks"
        items:
          $ref: "#/definitions/rule"
      validators:
        type: "array"
        description: "Object in order to validate the user answers"
        items:
          $ref: "#/definitions/validatorItem"
      flowConstraints:
        type: "array"
        description: "FlowConstraints to apply"
        items:
          $ref: "#/definitions/flowConstraint"
  SurveyImage:
    type: "object"
    properties:
      description:
        type: "string"
        description: "Description for the item"
      link:
        type: "string"
        description: "Link to the image file"
      labels:
        type: "array"
        description: "Objects that allow the surveyItem to be displayed in different\
          \ languages"
        items:
          $ref: "#/definitions/label"
      type:
        type: "string"
        description: "Name that identify the type of the surveyItemOption"
      options:
        type: "array"
        description: "Option that the user can select"
        items:
          $ref: "#/definitions/surveyItemOption"
      rules:
        type: "array"
        description: "Rules in order to apply consistency checks"
        items:
          $ref: "#/definitions/rule"
      validators:
        type: "array"
        description: "Object in order to validate the user answers"
        items:
          $ref: "#/definitions/validatorItem"
      flowConstraints:
        type: "array"
        description: "FlowConstraints to apply"
        items:
          $ref: "#/definitions/flowConstraint"
    example:
      validators:
      - id: 2
        type: "type"
        value: "value"
      - id: 2
        type: "type"
        value: "value"
      link: "link"
      options:
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      - id: 7
        type: "type"
        value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      description: "description"
      rules:
      - item: 3
        condition: "condition"
        id: 9
      - item: 3
        condition: "condition"
        id: 9
      type: "type"
      flowConstraints:
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      - typology: "typology"
        condition: "condition"
        id: 4
        value: "value"
      labels:
      - values: "values"
        language: "language"
        id: 2
      - values: "values"
        language: "language"
        id: 2
  body_5:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the SurveyItem"
      description:
        type: "string"
        description: "Description for the item"
      link:
        type: "string"
        description: "Link to the image file"
      labels:
        type: "array"
        description: "Objects that allow the surveyItem to be displayed in different\
          \ languages"
        items:
          $ref: "#/definitions/label"
      type:
        type: "string"
        description: "Name that identify the type of the surveyItemOption"
      options:
        type: "array"
        description: "Option that the user can select"
        items:
          $ref: "#/definitions/surveyItemOption"
      rules:
        type: "array"
        description: "Rules in order to apply consistency checks"
        items:
          $ref: "#/definitions/rule"
      validators:
        type: "array"
        description: "Object in order to validate the user answers"
        items:
          $ref: "#/definitions/validatorItem"
      flowConstraints:
        type: "array"
        description: "FlowConstraints to apply"
        items:
          $ref: "#/definitions/flowConstraint"
...
  validatorItem_1:
    type: "object"
    properties:
      type:
        type: "string"
        description: "Type of the validation value"
      value:
        type: "string"
        description: "Value in order to validate the data"
  body_8:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the validatorItem in the system"
      type:
        type: "string"
        description: "Type of the validation value"
      value:
        type: "string"
        description: "Value in order to validate the data"
...